Add testing support for http and json stub (#2)
* Add testing support for http and json stub

Move all http response fixtures and json fixtures in the lib so
these can be reused in testing the implementation in other ruby
applications or libraries.

* remove stray pry require

* Introduce `TestingSupport::HttpResponses`

for sharing the http and JSON fixtures so we can share these in environments that use this gem. refactored all the specs to use these helpers internally as well.